requirements.txt
gitdb2==2.0.5
GitPython==2.1.11
pkg-resources==0.0.0
PyYAML==5.1
smmap2==2.0.5
